+++ src/br/com/alura/humanDev/application/Program.java
@@ -1,0 +1,39 @@
+package br.com.alura.humanDev.application;
+
+import br.com.alura.humanDev.entities.*;
+import br.com.alura.humanDev.entities.activities.Explanation;
+import br.com.alura.humanDev.entities.activities.Question;
+import br.com.alura.humanDev.entities.activities.Video;
+
+public class Program {
+
+    public static void main(String[] args) {
+
+
+
+
+
+
+
+
+/*
+         Course course = new Course("Java", "codORRR3c2",3,  "mario");
+         Category category1 = new Category("Backend", "jhujfhHH77--kk");
+         Subcategory subcategory1 = new Subcategory("Variables", "jhf88hhd--d", category1 );
+         Section section1 = new Section("Study Section", "co9d3-aq", course);
+         Video video1 = new Video("Title", "co-d3e", section1, "http-00", 12);
+         Question question = new Question("Question Title", "co-od3", section1, "Description about the question");
+         Alternative alternative = new Alternative("This explanation is about the test", true, "the answer is A");
+         Explanation explanation = new Explanation("Explanation title", "cod3-", section1, "Activity explanation huge text");
+
+         System.out.println(course);
+         System.out.println(category1);
+         System.out.println(subcategory1);
+         System.out.println(section1);
+         System.out.println(video1);
+         System.out.println(question);
+         System.out.println(alternative);
+         System.out.println(explanation);
+*/
+    }
+}
