+++ src/br/com/alura/humanDev/entities/Subcategory.java
@@ -1,8 +1,8 @@
-package entities;
+package br.com.alura.humanDev.entities;
 
-import static validations.Validation.notNull;
-import static validations.Validation.notBlank;
-import static validations.CodePatternValidation.validUrl;
+import static br.com.alura.humanDev.validations.CodePatternValidation.validUrl;
+import static br.com.alura.humanDev.validations.Validation.notBlank;
+import static br.com.alura.humanDev.validations.Validation.notNull;
 
 public class Subcategory {
